+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog of lizard-nxt client
 Unreleased ()
 -------------
 
+- Force cursor behaviour in point, line and area mode.
+
 - Clean up dependencies.
 
 - Store global state in seperate module.